requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
requests==2.32.3
plexapi==4.16.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pycountry